bug: CreateEditView로 분기 (#61)
