+++ DietLog-MVVM/View/Exercise/Controller/ExerciseBaseEditViewController.swift
@@ -13,15 +13,15 @@ enum ExerciseEditOption: Int {
     case URL, category, memo
 }
 
-class ExerciseEditViewController: BaseViewController {
+class ExerciseBaseEditViewController: BaseViewController {
 
     // MARK: - Component
     lazy var exerciseEditView = ExerciseEditView()
     
     // MARK: - 변수
-    private var viewModel = ExerciseEditViewModel()
-    private var categoryViewModel = SelectCategoryViewModel()
-    private var disposeBag = DisposeBag()
+    var viewModel = ExerciseEditViewModel()
+    var categoryViewModel = SelectCategoryViewModel()
+    var disposeBag = DisposeBag()
 
     // MARK: - Life Cycle
     override func loadView() {
@@ -40,21 +40,8 @@ class ExerciseEditViewController: BaseViewController {
     
     // MARK: - Setup NavigationBar
     override func setupNavigationBar() {
-        let button = UIBarButtonItem(title: "저장", style: .plain, target: self, action: #selector(saveData))
-        navigationItem.rightBarButtonItem = button
-        
         let backButton = UIBarButtonItem(title: "", style: .plain, target: self, action: nil)
         navigationItem.backBarButtonItem = backButton
-    }
-    
-    @objc func saveData() {
-        let result = viewModel.saveData()
-        let message = result ? "저장했습니다" : "URL 입력 및 카테고리를 선택해 주세요"
-        self.showAlertWithOKButton(title: "", message: message) {
-            if result {
-                self.navigationController?.popViewController(animated: true)
-            }
-        }
     }
     
     // MARK: - Setup Bind
@@ -87,7 +74,7 @@ class ExerciseEditViewController: BaseViewController {
 }
 
 // MARK: - ExerciseEditView
-extension ExerciseEditViewController: ExerciseEditViewDelegate {
+extension ExerciseBaseEditViewController: ExerciseEditViewDelegate {
     func moveToSelectCategoryView() {
         let viewController = ExerciseSelectCategoryViewController(viewModel: categoryViewModel)
         navigationController?.pushViewController(viewController, animated: true)
