+++ DietLog-MVVM/View/Exercise/Controller/ExerciseModifyEditViewController.swift
@@ -1,0 +1,15 @@
+//
+//  ExerciseModifyEditViewController.swift
+//  DietLog-MVVM
+//
+//  Created by Jooyeon Kang on 2024/05/05.
+//
+
+import UIKit
+
+class ExerciseModifyEditViewController: ExerciseBaseEditViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
